pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "mnoisy"
dynamic = ["version"]
description = "Pseudorandom (m)noise generator from m sequences."
readme = "README.md"
requires-python = ">=3.8"
authors = [{ name = "Kevin Chu", email = "kvnchv@gmail.com" }]
dependencies = [
  # add requirements here
  "scipy",
  "matplotlib",
  "jsonschema",
  "pyyaml",
]

[project.scripts]
mnoisy-gen = "mnoisy:__main__.main"

[project.optional-dependencies]
test = ["pytest", "pytest-cov"]
dev = ["jupyter", "black", "isort", "ruff", "pre-commit"]

[tool.black]
line-length = 120
target-version = ['py310']


[tool.ruff]
# Exclude a variety of commonly ignored directories.
extend-exclude = ["examples/**/*.py", "other/**/*.py"]
line-length = 120
target-version = "py310"

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
